Get the PCI protocol working as far as the clock and reset.
The PCI_CLK and RESET# signals of the pci bus protocol now work from
the host device all the way through to the client devices. This includes
properly initializing the libsimbus pci device, implementing the protocol
for the clock and reset in the server, and implementing the drivers in
the pci_port.v and the simbus vpi code. Now the server has the support
for passing signal values around and the VPI has the code for receiving
driven values from the server.