Take a stab at handling setup/hold issues.
The server acts as a synchronous device in between the host and
the client, so get the host to shift its timing around so that the
client seems the signals with better timing. Also make sure the
non-clock signals are written with non-blocking assignments so that
they are scheduled correctly by the Verilog simulator.