Add server ability to trace (to LXT) the state of the bus.
The protocol handler of the server can trace not only the bus signals,
but can include some interesting bus state in the trace.

NOTE: This patch includes the lxt2 writer helper functions from the
gtkwave 3.1 source distribution. That is OK for now, but in the long
run it may be better to have the GTKWave distribution factored out
so that the libraries can be referenced as a package instead of being
included herin.