Subtract feedback when sending blended results to client
If the client is driving the signal, then do not feed back the value
that it is driving. This allows the client to know that it is indeed
itself that is doing the driving. If we do not subtract feedback, then
clients may start to believe that the value is owned by the bus cloud,
and chaos will reign.