Rework PCI protocol to have 4 phases of clock.
Giving clients hooks into 4 phases of the clock greatly
simplifies (and clarifies) the handling of signals around
clock edges. Since delta-times do not really exist in the
simbus protocol, the setup and hold phase points give a
clearer model of the operation of the bus. And the traces
look more natural when the server dumps them.