Make RequestImage edge sensitive / Fix implementation of -g flag.
The RequestImage is sampled (by the camera) on the rising edge
of the video clock, but the camera is still interested in the
rising edge of the request. Arm the capture when the request
is low, and initiate it when it goes high. This makes the timing
much easier to handle.

Also fix the implementation of the "-g" flag to the command.