Add the PCIeTLP tx_buf_av flow control, and some other stubs.
The tx_buf_av signal is really a Xilinx core thing, but the
xilinx_pcie_slot module doesn't really do buffering so pass
on the job of handling it to the root device, through the
server.

Also, implement or stub some of the more basic cfg_ signals for
the xilinx_pcie_slot module.